reformat index.js for facilitate testing
